+++ src/wave/begin.rs
@@ -1,4 +1,4 @@
-use crate::{debug, indent};
+use crate::{data::subskill::Trigger, debug, indent};
 
 use super::Wave;
 
@@ -6,6 +6,10 @@ impl Wave<'_> {
     pub fn begin(&mut self) {
         debug!("Wave begin");
         indent!({
+            for i in self.get_indices_iter() {
+                self.on_trigger(i, Trigger::WaveBegin);
+            }
+            // TODO these can be converted to subskills
             self.on_begin_wave_space();
             self.on_begin_wave_marville();
         })