+++ src/wave/heroes/asindo/mod.rs
@@ -1,0 +1,4 @@
+pub mod psychological_torture;
+
+use super::test_hero;
+test_hero!(Asindo);