+++ src/wave/turn_meter.rs
@@ -11,24 +11,40 @@ impl Wave<'_> {
         self.turn_meter[actor] = turn_meter;
     }
 
+    pub fn steal_turn_meter_ratio(
+        &mut self,
+        actor: InstanceIndex,
+        target: InstanceIndex,
+        turn_meter_ratio: f32,
+    ) {
+        let stolen = self.reduce_turn_meter_ratio(actor, target, turn_meter_ratio);
+        self.increase_turn_meter(actor, target, stolen)
+    }
+
     pub fn increase_turn_meter_ratio(
         &mut self,
         actor: InstanceIndex,
         target: InstanceIndex,
         turn_meter_ratio: f32,
     ) {
-        self.reduce_turn_meter(actor, target, turn_meter_ratio * self.turn_meter_threshold);
+        self.increase_turn_meter(actor, target, turn_meter_ratio * self.turn_meter_threshold);
     }
 
-    pub fn increase_turn_meter(&mut self, actor: InstanceIndex, turn_meter: f32) {
-        self.turn_meter[actor] = self
+    pub fn increase_turn_meter(
+        &mut self,
+        actor: InstanceIndex,
+        target: InstanceIndex,
+        turn_meter: f32,
+    ) {
+        self.turn_meter[target] = self
             .turn_meter_threshold
-            .min(self.turn_meter[actor] + turn_meter);
+            .min(self.turn_meter[target] + turn_meter);
         debug!(
-            "{} turn_meter increased by {} to {}",
+            "{} increases {}'s turn_meter by {} to {}",
             self.name(actor),
+            self.name(target),
             turn_meter,
-            self.turn_meter[actor]
+            self.turn_meter[target]
         );
     }
 
@@ -37,8 +53,8 @@ impl Wave<'_> {
         actor: InstanceIndex,
         target: InstanceIndex,
         turn_meter_ratio: f32,
-    ) {
-        self.reduce_turn_meter(actor, target, turn_meter_ratio * self.turn_meter_threshold);
+    ) -> f32 {
+        self.reduce_turn_meter(actor, target, turn_meter_ratio * self.turn_meter_threshold)
     }
 
     pub fn reduce_turn_meter(
@@ -46,7 +62,7 @@ impl Wave<'_> {
         actor: InstanceIndex,
         target: InstanceIndex,
         turn_meter: f32,
-    ) {
+    ) -> f32 {
         let turn_meter = turn_meter * (1.0 - self.get_turn_meter_reduction_reduction(target));
         debug!(
             "{} turn_meter reduced by {} from {} to {}",
@@ -55,7 +71,9 @@ impl Wave<'_> {
             self.turn_meter[target],
             self.turn_meter[target] - turn_meter
         );
+        let old = self.turn_meter[target];
         self.turn_meter[target] = (self.turn_meter[target] - turn_meter).max(0.0);
+        old - self.turn_meter[target]
     }
 
     pub fn progress_turn_meter(&mut self, actor: InstanceIndex, time: f32) {
@@ -71,7 +89,7 @@ impl Wave<'_> {
         debug!("{} increases turn meter of team", self.name(actor));
         indent!({
             self.get_ally_indices(actor).iter().for_each(|&i| {
-                self.increase_turn_meter(i, increase_ratio * self.turn_meter_threshold)
+                self.increase_turn_meter(actor, i, increase_ratio * self.turn_meter_threshold)
             })
         });
     }